Bump Kubevirt to version that supports Vm freeze (#20)
* Bump Kubevirt to newer version that supports Vm freeze

A latest Kubevirt v48.1 wich contains "Added virt-freezer
https://github.com/kubevirt/kubevirt/pull/6124" is used.

Signed-off-by: Bartosz Rybacki <brybacki@redhat.com>

* Bump Kubernetes - 1.21

Signed-off-by: Bartosz Rybacki <brybacki@redhat.com>

* Cleanup makefile target - go mod

Signed-off-by: Bartosz Rybacki <brybacki@redhat.com>

* Vendor in CDI 1.40

Vendor in newer CDI To get a NodePull method.

Signed-off-by: Bartosz Rybacki <brybacki@redhat.com>

* Use a vm with guest agent enabled

To correctly use freeze-unfreeze a guest agent has
to be installed in vm. Tests need to Wait for a
vm condition AgentConnected.

New VM image is bigger, so a NodePull method is used
to utilize node cache.

Signed-off-by: Bartosz Rybacki <brybacki@redhat.com>

* Do not fail in the AfterEach

Signed-off-by: Bartosz Rybacki <brybacki@redhat.com>

* Fix one and Exclude one test

One test is failing and needs to be investigated

Signed-off-by: Bartosz Rybacki <brybacki@redhat.com>