Switch to using podman, cleanup build logic (#72)
* Remvoe unused variables. Not used anywhere.

Signed-off-by: Maya Rashish <mrashish@redhat.com>

* Don't pre-pull fedora test image

We don't use it for testing. One less use of docker.

Signed-off-by: Maya Rashish <mrashish@redhat.com>

* Remove outdated reference to DOCKER=0 as a way to build on the host

Signed-off-by: Maya Rashish <mrashish@redhat.com>

* Stop passing unused flag CLUSTER_PREFIX

Signed-off-by: Maya Rashish <mrashish@redhat.com>

* Switch from docker to podman

Note that the kubevirtci docker is already podman, so there's
no change in the ssh.sh node01 code. It's just helpful to know
that we make no references to docker.

We don't relabel the directory to avoid permission errors on
rootless podman.

We don't validate tls when pushing to local repository, as it
doesn't have a valid certificate.

Signed-off-by: Maya Rashish <mrashish@redhat.com>

* Keep supporting docker and not just podman

Having podman on the CI has proven risky - it is breaking random
other CI lanes. We will transition gradually by supporting both
on all our repos & then moving CI images one by one.

Signed-off-by: Maya Rashish <mrashish@redhat.com>

* Don't change modules when building functional tests

Unclear why we do this, and it breaks the build.
It might be a problem introduced by using a newer image and Go version.

Signed-off-by: Maya Rashish <mrashish@redhat.com>

* Override KUBEVIRT_MEMORY_SIZE that is set by cluster-up

This previously didn't work because hack/cluster-up/common.sh is
run first. Change the order so we successfully override the size.

Signed-off-by: Maya Rashish <mrashish@redhat.com>