Do not use snapshot location in tests
Snapshot location is actually not needed in the tests. Snapshot location
can and should be configured when running on a storage provider that
has an option to store snapshot in configurable locations. This is not
possible on ceph provider used in functional tests, which stores snapshots
in the same location as normal volumes.

Signed-off-by: Bartosz Rybacki <brybacki@redhat.com>