Backup VM and VMI object graph (#124)
* Backup up all volumes when backing up a vm

In case the vm is not running then we will not
backup up volumes which are not on the datavolumetemplate
This should fix that.

Signed-off-by: Shelly Kagan <skagan@redhat.com>

* Add datavolumetemplate to the example vm yaml

Signed-off-by: Shelly Kagan <skagan@redhat.com>

* Move create started vm to tests framework

Signed-off-by: Shelly Kagan <skagan@redhat.com>

* Add tests for vm backup with standalone dv

Signed-off-by: Shelly Kagan <skagan@redhat.com>

* Backup VM and VMI object graph

In case the backup that occures is only for label for example
which backs up only the vm/vmi object we want all the related objects
in the graph to also be included in the backup.

Signed-off-by: Shelly Kagan <skagan@redhat.com>

* Bump kubevirt version to 0.58

Signed-off-by: Shelly Kagan <skagan@redhat.com>

* Add tests for vm and vmi object graph backup

Signed-off-by: Shelly Kagan <skagan@redhat.com>

Signed-off-by: Shelly Kagan <skagan@redhat.com>