Test and env updates (#208)
* Add PartnerComp label

This label can be used to run tests with focus
on tests we consider neccessary to test partner
compatibility with kubevirt.
To run the tests:
make TEST_ARGS="--test-args=-ginkgo.label-filter=PartnerComp" test-functional

Signed-off-by: Shelly Kagan <skagan@redhat.com>

* Adjust tests to have configurable storageclass yamls

This change allows the user to set env variable
KVP_STORAGE_CLASS with the desired storage class
one wants the tests applying yamls to run with.
Adjust some tests to pass when using wffc SC

Signed-off-by: Shelly Kagan <skagan@redhat.com>

* Fix racy test

The test wants to catch the dv in state importing
but if the import is fast we can miss it.
Insted we can check that a file that existed
no longer exists after the backup.

Signed-off-by: Shelly Kagan <skagan@redhat.com>

* Ignore velero files in gitignore, delete existing

Signed-off-by: Shelly Kagan <skagan@redhat.com>

* Update kubevirtci due to a bug in last version

Signed-off-by: Shelly Kagan <skagan@redhat.com>

---------

Signed-off-by: Shelly Kagan <skagan@redhat.com>