Clear MAC address by label
This commit allows to optionally clear the MAC address in VM and VMIs during restore.

Currently, the way to go for modular backup/restore behavior in velero plugin is by using labels instead of annotations.

Labels are easier to add to the restore/backup objects during creation as the velero client has a flag to include them.

This commit also changes some const names to use PascalCase to match exported const naming standard.

Signed-off-by: Eduardo Esteban <eesteban@live.com>

Signed-off-by: Alvaro Romero <alromero@redhat.com>