Fix cluster-sync due to velero 1.14 changes
* As stated in velero 1.14 release docs:
`CSI plugin has been merged into velero repo in v1.14 release. It will be installed by default as an internal plugin, and should not be installed via "–plugins " parameter in "velero install" command.`
As such removed the installation of velero-plugin-for-csi
* Velero added a confirmation check when adding a plugin when
the velero pod is already running, hence needed to add the `confirm`
flag.
* Increased pod memory request and limit due to OOMKilled with smaller
amount.
* updated aws plugin to v1.10 version

Signed-off-by: Shelly Kagan <skagan@redhat.com>