Bump velero library and deployment to 1.14 + go update and kubevirtci (#249)
* Update velero library and deployment version to 1.14

Also bump go version and builder to 1.22.3,
kubevirtci version and kubevirt provider to 1.29

Signed-off-by: Shelly Kagan <skagan@redhat.com>

* Run make gomod-update

Signed-off-by: Shelly Kagan <skagan@redhat.com>

* Use crictl instead of docker to pull from quay

https://github.com/kubevirt/kubevirtci/pull/1100

Signed-off-by: Shelly Kagan <skagan@redhat.com>

* Fix cluster-sync due to velero 1.14 changes

* As stated in velero 1.14 release docs:
`CSI plugin has been merged into velero repo in v1.14 release. It will be installed by default as an internal plugin, and should not be installed via "–plugins " parameter in "velero install" command.`
As such removed the installation of velero-plugin-for-csi
* Velero added a confirmation check when adding a plugin when
the velero pod is already running, hence needed to add the `confirm`
flag.
* Increased pod memory request and limit due to OOMKilled with smaller
amount.
* updated aws plugin to v1.10 version

Signed-off-by: Shelly Kagan <skagan@redhat.com>

* Update ginkgo cli in tests to 2.17.1

Signed-off-by: Shelly Kagan <skagan@redhat.com>

---------

Signed-off-by: Shelly Kagan <skagan@redhat.com>