Implement KVGraph package to allow retrieving a defined list of vm-related objects (#260)
* Create a exportable VM object graph

This commit creates a new vmgraph package that allows third-party usage of our vm object graph

Signed-off-by: Alvaro Romero <alromero@redhat.com>

* Add virt-launcher pod into the VM and VMI object graphs

Signed-off-by: Alvaro Romero <alromero@redhat.com>

* Implement restore object graph

This commit rearranges the vmgraph package so we have backup and restore-specific functions and files.

Signed-off-by: Alvaro Romero <alromero@redhat.com>

* Add error handling in backup graph

Signed-off-by: Alvaro Romero <alromero@redhat.com>

* Implement DataVolume backup graph

This commit implements a function to get the DataVolume backup graph and adds proper test coverage.

Signed-off-by: Alvaro Romero <alromero@redhat.com>

* Add a generic function to get the object graph for any supported item

This commit adds a common function for all object kinds so we can fetch the full object graph for different items using the same function.

This allows us to easily maintain and update the resources we backup/restore without the need for external users to learn about new functions.

Signed-off-by: Alvaro Romero <alromero@redhat.com>

* Rename vmgraph package to kvgraph

Signed-off-by: Alvaro Romero <alromero@redhat.com>

---------

Signed-off-by: Alvaro Romero <alromero@redhat.com>