[release-v0.7] Ignore VMIs from backup when owner VMs are excluded (#315)
* Fix VMI backups when VMs are excluded

This commit updates the backup process for VMIs so, when the VMI is owned by a VM that's excluded from the backup, we also exclude the VMI without triggering an error.

Signed-off-by: Alvaro Romero <alromero@redhat.com>

* Update functional tests

This commit updates the functional tests to match the new behavior where we ignore the VMI backup if its owner is excluded from the backup.

Signed-off-by: Alvaro Romero <alromero@redhat.com>

---------

Signed-off-by: Alvaro Romero <alromero@redhat.com>