[release-v0.7] VM object graph (#324)
* Create a exportable VM object graph

This commit creates a new vmgraph package that allows third-party usage of our vm object graph

Signed-off-by: Alvaro Romero <alromero@redhat.com>

* Add virt-launcher pod into the VM and VMI object graphs

Signed-off-by: Alvaro Romero <alromero@redhat.com>

* Implement restore object graph

This commit rearranges the vmgraph package so we have backup and restore-specific functions and files.

Signed-off-by: Alvaro Romero <alromero@redhat.com>

* Add error handling in backup graph

Signed-off-by: Alvaro Romero <alromero@redhat.com>

* Implement DataVolume backup graph

This commit implements a function to get the DataVolume backup graph and adds proper test coverage.

Signed-off-by: Alvaro Romero <alromero@redhat.com>

* Add a generic function to get the object graph for any supported item

This commit adds a common function for all object kinds so we can fetch the full object graph for different items using the same function.

This allows us to easily maintain and update the resources we backup/restore without the need for external users to learn about new functions.

Signed-off-by: Alvaro Romero <alromero@redhat.com>

* Rename vmgraph package to kvgraph

Signed-off-by: Alvaro Romero <alromero@redhat.com>

* Add persistent state PVC into VM object graph

A VM might use a backend PVC to persistently store its state. This volume is not included in the VM's spec, thus being ignored by our object graph.

This commit updates the VM object graph by including this PVC if necessary.

Signed-off-by: Alvaro Romero <alromero@redhat.com>

* Add func test coverage for persistent state PVC

Signed-off-by: Alvaro Romero <alromero@redhat.com>

* Update deployment scripts to support testing of backend volumes

This commit updates the following:

* Enables VMPersistentState feature gate in kubevirt config.
* Sets vmStateStorageClass to the default storage class in kubevirt config.

This fulfill the prerequisites so our func tests can work with VMs using persistent backend PVCs.

Signed-off-by: Alvaro Romero <alromero@redhat.com>

* Bump Kubevirt version from v1.1.1 to v1.3.0

This change is needed so we can create backend PVCs without RWX access mode.

Signed-off-by: Alvaro Romero <alromero@redhat.com>

---------

Signed-off-by: Alvaro Romero <alromero@redhat.com>