Do not skip PVC when building VMIs resources graph during restore item action.
In order to restore one or more Virtual Machines out of many in the SAME namespace, velero offers restoring by resource label.
When using a label selector, any resource not including the restore label are excluded by velero.
That might include other resources that are quieted for the VM, such as datavolumes and PVCs.
In order to resolve this, we already include any additional required resource in a graph which is return to velero for backup/restore even if those resources do not include matching label.

With this PR we are removing a condition results in excluding PVC for datavolume type VM volume from the graph, and a the datavolume pending for ever for the existence of the associated PVC after restore.

Signed-off-by: Amos Mastbaum <amastbau@redhat.com>